requirements.txt
accelerate==0.34.2
bitsandbytes==0.42.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==7.1.0
dill==0.3.8
diskcache==5.6.3
etelemetry==0.3.1
faiss-cpu==1.8.0.post1
filelock==3.15.4
fitz==0.0.1.dev2
Flask==3.0.3
fsspec==2024.9.0
httplib2==0.22.0
huggingface-hub==0.24.6
idna==3.8
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
llama_cpp_python==0.2.90
looseversion==1.3.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multiprocess==0.70.16
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
nltk==3.9.1
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pathlib==1.0.1
pillow==10.4.0
prov==2.0.1
psutil==6.0.0
pydot==3.0.1
Pygments==2.18.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pyparsing==3.1.4
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.2
rdflib==6.3.2
regex==2024.7.24
requests==2.32.3
rich==13.8.0
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.14.1
semchunk==2.2.0
sentence-transformers==3.0.1
sentencepiece==0.2.0
simplejson==3.19.3
six==1.16.0
sympy==1.13.2
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.66.5
traits==6.3.2
transformers==4.44.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.4
